ytdl_drive/requirements.txt
youtube-dl@git+https://github.com/ytdl-org/youtube-dl.git
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
